pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["YehorSaz <inercoin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
django-filter = "^23.3"
mysqlclient = "^2.2.0"
djangorestframework-simplejwt = "^5.3.0"
pillow = "^10.1.0"
drf-yasg = "^1.21.7"
redis = "^5.0.1"
celery = "^5.3.6"
django-celery-beat = "^2.5.0"
django-celery-results = "^2.5.1"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
django-dotenv = "^1.4.2"
isort = "^5.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
